Break down FirestoreRepository into subclasses
FirestoreRepository was too big, so it had to be broken down into the
two interfaces (HabitRepository and HabitEventRepository)

Also renamed `firestore` to `firebase` to include the `auth` package
inside.
