+++ app/src/main/java/com/github/cmput301f21t44/hellohabits/firebase/FirebaseTask.java
@@ -1,0 +1,11 @@
+package com.github.cmput301f21t44.hellohabits.firebase;
+
+public interface FirebaseTask {
+    interface ThenFunction {
+        void apply();
+    }
+
+    interface CatchFunction {
+        void apply(Exception e);
+    }
+}
